Issue #214: '"Output format" in SEARCH SETTINGS doesn't do anything'
Fixed at least partially by setting visibility to 1 for the default
configuration where no collections is in use. It is not clear how/if
b2share is going to use formats specific per search collection so
further fixes may be needed. If so, start with checking the database
tables collection_format, collection and format.